Added 71impdata test, - changing tests to error our if can't connect - took out code that loops through to include lib.pl when simple use lib/require suffices
git-svn-id: https://svn.perl.org/modules/DBD-mysql/trunk@11150 50811bd7-b8ce-0310-adc1-d9db26280581

+++ t/20createdrop.t
@@ -1,25 +1,24 @@
 #!perl -w
 # vim: ft=perl
 
-use Test::More tests => 4;
+use Test::More;
 use DBI;
 use DBI::Const::GetInfoType;
 use strict;
 $|= 1;
 
-my $mdriver= "";
-our ($test_dsn, $test_user, $test_password);
-foreach my $file ("lib.pl", "t/lib.pl") {
-  do $file;
-  if ($@) {
-    print STDERR "Error while executing $file: $@\n";
-    exit 10;
-  }
-  last if $mdriver ne '';
-}
+our ($mdriver, $test_dsn, $test_user, $test_password);
+use lib 't', '.';
+require 'lib.pl';
 
 my $dbh= DBI->connect($test_dsn, $test_user, $test_password,
                       { RaiseError => 1, PrintError => 1, AutoCommit => 0 });
+
+if (! defined $dbh) {
+    plan skip_all => "Can't obtain driver handle. Can't continue test";
+}
+plan tests => 4;
+
 ok(defined $dbh, "Connected to database");
 
 ok($dbh->do(qq{DROP TABLE IF EXISTS t1}), "making slate clean");
