+++ t/35limit.t
@@ -1,7 +1,7 @@
 #!perl -w
 # vim: ft=perl
 
-use Test::More tests => 111;
+use Test::More;
 use DBI;
 use DBI::Const::GetInfoType;
 use strict;
@@ -12,17 +12,15 @@ my $sth;
 my $testInsertVals;
 our ($test_dsn, $test_user, $test_password, $mdriver);
 $mdriver='';
-foreach my $file ("lib.pl", "t/lib.pl") {
-  do $file;
-  if ($@) {
-    print STDERR "Error while executing $file: $@\n";
-    exit 10;
-  }
-  last if $mdriver ne '';
-}
+use lib 't', '.';
+require 'lib.pl';
 
 my $dbh= DBI->connect($test_dsn, $test_user, $test_password,
                       { RaiseError => 1, PrintError => 1, AutoCommit => 0 });
+if (! defined $dbh) {
+    plan skip_all => "Can't connect to database ERROR: $DBI::errstr. Can't continue test";
+}
+plan tests => 111; 
 ok(defined $dbh, "Connected to database");
 
 ok($dbh->do(qq{DROP TABLE IF EXISTS t1}), "making slate clean");
